feat: add module createRoom
