+++ src/ws/modules/messageHandlers.ts
@@ -1,6 +1,8 @@
 import GamesRoom from "../db/GameRooms";
-import SessionDB, { SessionDBType } from "../db/SessionDB";
-import UsersDB, { UserDBType } from "../db/UsersDB";
+import { SessionDBType } from "../db/SessionDB";
+import UsersDB from "../db/UsersDB";
+import { connectUserToRoom } from "./connectUserToRoom/connectUserToRoom";
+import { createRoom } from "./createRoom/createRoom";
 import reg from "./reg/reg";
 
 export type DataMessage = {
@@ -20,8 +22,6 @@ const dbRoom = new GamesRoom();
 export default function messageHandlers(message: string, props: PropsUsers) {
   const { dbSession, sessionId } = props;
 
-  const currentUser = dbSession.getUserSession(sessionId);
-
   const parsedMessage = JSON.parse(message);
 
   let { data, type, id } = parsedMessage as DataMessage;
@@ -40,103 +40,24 @@ export default function messageHandlers(message: string, props: PropsUsers) {
       });
       break;
     case "create_room":
-      const newRoomId = dbRoom.createRoom();
-
-      const user = dbUser.getPlayerByLogin(currentUser.name);
-
-      if (user) {
-        dbRoom.addUserToRoom(newRoomId, {
-          index: user.id,
-          name: user.name,
-          sessionId,
-        });
-
-        const rooms = dbRoom.getAllRooms().map((room) => ({
-          roomId: room.idGame,
-          roomUsers: [
-            {
-              name: room.user1.name,
-              index: room.user1.index,
-            },
-          ],
-        }));
-
-        const dataUpdateRoom = JSON.stringify(rooms);
-
-        const responseUpdateRoom = JSON.stringify({
-          type: "update_room",
-          data: dataUpdateRoom,
-        });
-
-        Object.values(dbSession.getAllSessions()).forEach((session) => {
-          session.ws.send(responseUpdateRoom);
-        });
-
-        console.log(`Rooms update!`);
-      }
+      createRoom({
+        dbRoom,
+        dbSession,
+        dbUser,
+        sessionId,
+      });
       break;
     case "add_user_to_room":
-      const room = dbRoom.getRoomByIndex(data.indexRoom);
-
-      if (room?.user1.name !== currentUser.name) {
-        const result = dbRoom.connectUserToRoom(data.indexRoom, {
-          index: currentUser.id,
-          name: currentUser.name,
-          sessionId: currentUser.sessionId,
-        });
-
-        const rooms = dbRoom.getAllRooms().map((room) => ({
-          roomId: room.idGame,
-          roomUsers: [
-            {
-              name: room.user1.name,
-              index: room.user1.index,
-            },
-            {
-              name: room.user2.name,
-              index: room.user2.index,
-            },
-          ],
-        }));
-
-        const dataUpdateRoom = JSON.stringify(rooms);
-
-        const responseUpdateRoom = JSON.stringify({
-          type: "update_room",
-          data: dataUpdateRoom,
-        });
-
-        Object.values(dbSession.getAllSessions()).forEach((session) => {
-          session.ws.send(responseUpdateRoom);
-
-          if (
-            session.id === room?.user1.index ||
-            session.id === room?.user2.index
-          ) {
-            const dataGame = JSON.stringify({
-              idGame: room.idGame,
-              idPlayer: session.id,
-            });
-
-            const responseCreateGame = JSON.stringify({
-              type: "create_game",
-              data: dataGame,
-              id,
-            });
-
-            session.ws.send(responseCreateGame);
-
-            console.log(`User ${session.name} started the game!`);
-          }
-        });
-        console.log(result);
-      } else {
-        console.log("You can not join your room.");
-      }
+      connectUserToRoom({
+        dbRoom,
+        dbSession,
+        parsedMessage: { data, type, id },
+        sessionId
+      })
       break;
     case "add_ships":
       console.log(type);
-      console.log(type);
+      console.log(data);
       break;
     case "attack":
       console.log(type);
